TRX _BOT_GAME/requirements.txt
aiogram~=2.25.1
gino~=1.0.1
SQLAlchemy~=1.3.24
pandas~=2.0.1
python-dotenv~=1.0.0
openpyxl~=3.1.2
APScheduler~=3.9.1.post1